requirements.txt
easyocr==1.7.1
ultralytics==8.0.196
label-studio==1.12.1
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
numpy==1.24.3
streamlit==1.27.2
scikit-learn==1.3.1
PyYAML==6.0.1
flask==3.0.0
werkzeug==3.0.1
